fix some u64 bugs to support the 64 bits system

+++ include/common/mm.h
@@ -21,17 +21,17 @@
 #define L2_INDEX(addr) ((addr >> 21) & mask_9_bit)
 #define L3_INDEX(addr) ((addr >> 12) & mask_9_bit)
 
-#define L0_entry_addr(entry)      ((entry.paddr) << 12)
-#define L1_entry_huge_addr(entry) ((entry.paddr) << 30)
-#define L1_entry_addr(entry)      ((entry.paddr) << 12)
-#define L2_entry_huge_addr(entry) ((entry.paddr) << 21)
-#define L2_entry_addr(entry)      ((entry.paddr) << 12)
-#define L3_entry_addr(entry)      ((entry.paddr) << 12)
+#define L0_entry_addr(entry)      (((u64)(entry.paddr)) << 12)
+#define L1_entry_huge_addr(entry) (((u64)(entry.paddr)) << 30)
+#define L1_entry_addr(entry)      (((u64)(entry.paddr)) << 12)
+#define L2_entry_huge_addr(entry) (((u64)(entry.paddr)) << 21)
+#define L2_entry_addr(entry)      (((u64)(entry.paddr)) << 12)
+#define L3_entry_addr(entry)      (((u64)(entry.paddr)) << 12)
 
 #define PPN(paddr) (paddr >> 12)
 #define VPN(vaddr) (vaddr >> 12)
-#define PADDR(ppn) (ppn << 12)
-#define VADDR(vpn) (vpn << 12)
+#define PADDR(ppn) ((u64)ppn << 12)
+#define VADDR(vpn) ((u64)vpn << 12)
 
 enum ENTRY_FLAGS {
         PAGE_ENTRY_NONE = 1 << 0,
